cleaning the code
str_function -> get_verticapy_function
nearest_column -> self.get_nearest_column
sort_str -> self.__get_sort_syntax__

current_conn and current_cursor are now in connection.py
load_dataset_num, load_dataset_reg, load_dataset_cl used to simplify the tests
XGBoost_utils to simplify XGBoost methods usage
get_tree_list_of_arrays to simplify the tree methods
vModel.get_vertica_param_dict
cleaning toolbox
type_code_dtype -> type_code_to_dtype
