+++ verticapy/tests/vModel/test_countvectorizer.py
@@ -19,9 +19,9 @@ from verticapy import (
     drop,
     set_option,
     vertica_conn,
-    current_cursor,
     create_verticapy_schema,
 )
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic
 from verticapy.learn.preprocessing import CountVectorizer
