+++ verticapy/tests/vModel/test_elastic_net.py
@@ -21,8 +21,9 @@ import warnings
 import matplotlib.pyplot as plt
 
 # VerticaPy
+from verticapy import drop, set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_winequality
-from verticapy import drop, set_option, current_cursor
 from verticapy.learn.linear_model import ElasticNet
 
 set_option("print_info", False)
