+++ verticapy/tests/vModel/test_kmeans.py
@@ -18,7 +18,8 @@ import pytest
 import matplotlib.pyplot as plt
 
 # VerticaPy
-from verticapy import drop, set_option, current_cursor
+from verticapy import drop, set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_iris, load_winequality
 from verticapy.learn.cluster import KMeans
