+++ verticapy/tests/vModel/test_knn_classifier.py
@@ -22,8 +22,8 @@ from verticapy import (
     drop,
     set_option,
     create_verticapy_schema,
-    current_cursor,
 )
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic
 from verticapy.learn.neighbors import KNeighborsClassifier
