+++ verticapy/tests/vModel/test_logistic_regression.py
@@ -18,7 +18,8 @@ import pytest
 import matplotlib.pyplot as plt
 
 # VerticaPy
-from verticapy import drop, set_option, current_cursor
+from verticapy import drop, set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_winequality, load_titanic
 from verticapy.learn.linear_model import LogisticRegression
