+++ verticapy/tests/vModel/test_model_selection.py
@@ -19,7 +19,8 @@ import matplotlib.pyplot as plt
 
 # VerticaPy
 import verticapy
-from verticapy import vDataFrame, set_option, current_cursor
+from verticapy import vDataFrame, set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic, load_amazon, load_winequality
 from verticapy.learn.model_selection import *
 from verticapy.learn.linear_model import *
