+++ verticapy/tests/vModel/test_randomforest_classifier.py
@@ -22,10 +22,9 @@ from verticapy import (
     vDataFrame,
     drop,
     set_option,
-    current_cursor,
-    dataset_cl,
 )
-from verticapy.datasets import load_titanic
+from verticapy.connect import current_cursor
+from verticapy.datasets import load_titanic, load_dataset_cl
 from verticapy.learn.ensemble import RandomForestClassifier
 
 set_option("print_info", False)
@@ -33,7 +32,7 @@ set_option("print_info", False)
 
 @pytest.fixture(scope="module")
 def rfc_data_vd():
-    rfc_data_vd = dataset_cl(table_name="rfc_data", schema="public")
+    rfc_data_vd = load_dataset_cl(table_name="rfc_data", schema="public")
     yield rfc_data_vd
     drop(name="public.rfc_data", method="table")
