+++ verticapy/tests/vModel/test_svd.py
@@ -15,7 +15,8 @@
 import pytest
 
 # VerticaPy
-from verticapy import drop, set_option, current_cursor
+from verticapy import drop, set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_winequality
 from verticapy.learn.decomposition import SVD
