+++ verticapy/tests/vModel/test_tools.py
@@ -18,7 +18,8 @@ import pytest
 import matplotlib.pyplot as plt
 
 # VerticaPy
-from verticapy import set_option, current_cursor
+from verticapy import set_option
+from verticapy.connect import current_cursor
 from verticapy.datasets import load_titanic
 from verticapy.learn.model_selection import *
 from verticapy.learn.linear_model import *
